+++ src/main/kotlin/id/walt/vclib/Defaults.kt
@@ -23,6 +23,7 @@ object Defaults {
             register<GaiaxServiceOffering>(GaiaxServiceOffering)
             register<VerifiableVaccinationCertificate>(VerifiableVaccinationCertificate)
             register<ProofOfResidence>(ProofOfResidence)
+            register<VerifiableMandate>(VerifiableMandate)
         }
     }